+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* sqlalchemy's DATETIME type now rendered as YDB's Datetime instead of Timestamp
+
 ## 0.0.1b19 ##
 * Do not use set for columns in index, use dict (preserve order)